loofyloo_prime/data.py

import os
import torch
import librosa
from PIL import Image
from torch.utils.data import Dataset, DataLoader
from transformers import BertTokenizer
from torchvision import transforms

# This is a skeleton implementation of a multimodal dataset.
# To make this a functional data loader, you would need to implement
# the __getitem__ and __len__ methods to load and preprocess your
# text, image, and audio data.

class MultimodalDataset(Dataset):
    """
    A multimodal dataset that loads text, image, and audio files.
    """
    def __init__(self, data_dir, embed_dim):
        """
        Initializes the MultimodalDataset.

        Args:
            data_dir (str): The directory containing the data.
        """
        self.data_dir = data_dir
        self.embed_dim = embed_dim
        self.text_files = os.listdir(os.path.join(data_dir, "text"))
        self.image_files = os.listdir(os.path.join(data_dir, "image"))
        self.audio_files = os.listdir(os.path.join(data_dir, "audio"))
        self.tokenizer = BertTokenizer.from_pretrained("bert-base-uncased")
        self.image_transform = transforms.Compose([
            transforms.Resize((224, 224)),
            transforms.ToTensor(),
        ])

    def __getitem__(self, idx):
        """
        Returns a single sample of data.

        Args:
            idx (int): The index of the sample to return.

        Returns:
            dict: A dictionary containing the text, image, and audio data.
        """
        text_path = os.path.join(self.data_dir, "text", self.text_files[idx])
        with open(text_path, "r") as f:
            text = f.read()
        text = self.tokenizer(text, return_tensors="pt", padding="max_length", max_length=128)["input_ids"].squeeze(0)

        image_path = os.path.join(self.data_dir, "image", self.image_files[idx])
        image = Image.open(image_path).convert("RGB")
        image = self.image_transform(image)

        audio_path = os.path.join(self.data_dir, "audio", self.audio_files[idx])
        audio, _ = librosa.load(audio_path, sr=16000)
        audio = torch.tensor(audio).float()

        return {
            "text": text,
            "image": image,
            "audio": audio,
            "target": torch.randn(128, self.embed_dim),
        }

    def __len__(self):
        """
        Returns the total number of samples in the dataset.

        Returns:
            int: The total number of samples in the dataset.
        """
        return len(self.text_files)

def get_data_loader(data_dir, batch_size, embed_dim):
    """
    Returns a DataLoader for the MultimodalDataset.

    Args:
        data_dir (str): The directory containing the data.
        batch_size (int): The batch size.

    Returns:
        DataLoader: A DataLoader for the MultimodalDataset.
    """
    dataset = MultimodalDataset(data_dir, embed_dim)
    return DataLoader(dataset, batch_size=batch_size)
