requirements.txt
torch
torchvision
transformers
librosa
Pillow
PyYAML
